requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.11.0
django-heroku==0.3.1
django-environ==0.8.1
psycopg2==2.9.3
sqlparse==0.4.2
whitenoise==5.3.0
djangorestframework==3.13.1
Pillow==9.0.0